Suppress the numerous icc11 compiler remarks and warnings.
git-svn-id: http://svn.cern.ch/guest/lhcb/Analysis/trunk@137887 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 80046ae7ea5b494f2f7028b29d06ca9e863074f0

+++ Analysis/Phys/DecayTreeTuple/src/TupleToolBase.h
@@ -65,7 +65,7 @@ protected:
   template<class CLASS>
   inline int number( const std::string& location)
   {
-    return ( safeExist<CLASS>(location) ? (get<CLASS>(location))->size() : -1 );
+    return ( safeExist<CLASS>(location) ? (int)(get<CLASS>(location))->size() : -1 );
   }
 
   /// Get the TES location of an object.
